PERF factor out camel_to_snake to its own module
