+++ src/civis/_utils.py
@@ -16,8 +16,7 @@ from tenacity.wait import wait_base
 
 
 log = logging.getLogger(__name__)
-UNDERSCORER1 = re.compile(r"(.)([A-Z][a-z]+)")
-UNDERSCORER2 = re.compile("([a-z0-9])([A-Z])")
+
 MAX_RETRIES = 10
 
 _RETRY_CODES = [429, 502, 503, 504]
@@ -29,12 +28,6 @@ def maybe_get_random_name(name):
     if not name:
         name = uuid.uuid4().hex
     return name
-
-
-def camel_to_snake(word):
-    # https://gist.github.com/jaytaylor/3660565
-    word = UNDERSCORER1.sub(r"\1_\2", word)
-    return UNDERSCORER2.sub(r"\1_\2", word).lower()
 
 
 def to_camelcase(s):
